gatekeeper/testing: testing package
Introduce a testing package which introduces a base mock type which can
be used for building and asserting against mocks locally.

Future iterations of this type will build out base mocks that can be
used for other tests and such. For instance, a MockServiceContainer will
be created if needed.

Introduce assertions that can be used throughout the codebase.