gatekeeper/utils: config package for parsing plugin config
Build out a package for parsing configuration of flags passed from the
gatekeeper core application into various plugins.

Behind the scenes, this package uses reflection to inspect the config
type for tags and then attempst to coerce raw values (either string or
bool) into said type.

Support for a `default` tag is enabled, allowing a default value to be
used in the case of a missing value.